run QC diagnostics for WC or SMRU tag workflows

produces a map of all QC'd tracks and generates various diagnostics to assess QC run

diagnostics(
  fit,
  fit1,
  what = "p",
  cut,
  data,
  ssm,
  meta,
  lines = FALSE,
  obs = FALSE,
  mpath = NULL,
  dpath = NULL,
  QCmode = "nrt",
  tag_mfr = "wc",
  cid = NULL
)

Arguments

fit

the final aniMotum fit object from QC process

fit1

the initial aniMotum fit object from QC process

what

the SSM-estimated or rerouted locations to be used

cut

logical; should predicted locations be dropped if keep = FALSE - ie. in a large data gap

data

the standardized WC Locations or SMRU diag file (prior to truncation by metadata CTD start and end dates)

ssm

the ssm-annotated WC/SMRU tables

meta

metadata

lines

add track lines to map (default = FALSE)

obs

add observed locations to map (default = FALSE)

mpath

path to write map file

dpath

path to write all other diagnostic files

QCmode

specify whether QC is near real-time (nrt) or delayed-mode (dm), in latter case start end end of dive data are displayed rather than ctd data.

tag_mfr

the tag manufacturer. Currently, only smru or wc are supported.

cid

SMRU campaign id (from config file). Ignored if WC data is used.