Download SMRU or Wildlife Computers tag data from their servers

Satellite tracking data are accessed from the SMRU data server, or accessed from the Wildlife Computers Portal API via the source argument.

download_data(
  dest = NULL,
  source = "smru",
  cid = NULL,
  user = NULL,
  pwd = NULL,
  wc.akey = NULL,
  wc.skey = NULL,
  owner.id = NULL,
  subset.ids = NULL,
  download = TRUE,
  ...
)

Arguments

dest

destination path to save download

source

source type of data to be downloaded. Can be one of:

smru - SMRU Data Server;

wc - Wildlife Computers Data Portal API;

cid

SMRU tag deployment campaign id(s) to download, eg. "ct180"

user

SMRU data server username as a quoted string

pwd

SMRU data server password as a quoted string

wc.akey

an Access Key issued by Wildlife Computers for their API

wc.skey

a Secret Key issued by Wildlife Computers for their API

owner.id

the Wildlife Computers uuid associated with the data owner

subset.ids

a single column .CSV file of WC UUID's to be included in the QC, with uuid as the variable name.

download

(logical) indicating if the data is to be downloaded from the tag manufacturer's server. If the source is wc and download = FALSE then only the Wildlife Computers tag deployment metadata is downloaded.

...

additional arguments passed to smru_get_mdb() or wc_get_files()