download deployment metadata

downloads, restructures & formats metadata, appends dive/CTD start and end datetimes (for QC), & fills in missing required metadata - eg. release_datetime, release_longitude/latitude's with data from the GPS (if present) or Argos location file.

get_metadata(
  source = "smru",
  tag_mfr = "smru",
  tag_data = NULL,
  cid = NULL,
  user = NULL,
  pwd = NULL,
  dropIDs = NULL,
  file = NULL,
  meta.args,
  subset.ids = NULL,
  wc.meta = NULL
)

Arguments

source

the source of the deployment metadata, current options are imos, smru, or atn. If source = 'imos' or source = 'atn' then metadata are obtained from user-provided .CSV file via the config.json file. If sourcce = 'smru' then metadata are built from a combination of SMRU server details & deployment details in the config.json file.

tag_mfr

the tag manufacturer, current options are smru or wc cid SMRU campaign ids

tag_data

a list of either smru data tables or wc files as output by pull_data.

cid

SMRU campaign id must be provided when the tag_mfr is smru

user

SMRU data server username as a quoted string - to be used only if metadata are to be built from SMRU server details (source = 'smru').

pwd

SMRU data server password as a quoted string - to be used only if metadata are to be built from SMRU server details (source = 'smru').

dropIDs

SMRU refs or WC ids to be dropped from QC

file

path to metadata .csv file, if provided then metadata will be read from the provided source

meta.args

optional metadata fields to be passed from config file when downloading tag metadata from SMRU server. Typically used only when no metadata filepath is provided in the config file.

subset.ids

a character vector of comma-separated (no spaces) WC UUID's to be included in the QC. Ignored if tag_mfr != "wc".

wc.meta

an R data.frame of Wildlife Computers tag deployment metadata obtained via download_data(). Typically used only when no metadata filepath is provided in the config file.