Pull SMRU or WC data from downloaded data files

extracts specified data files from SMRU .mdb files or WC zipfiles. For WC data files, extracts data from -Locations.csv, -FastGPS.csv, Histos.csv, MinMaxDepth.csv, HaulOut.csv, and SST.csv files. Extracted data are aggregated across individual tags and returned in a single named list with the following data.frames:

argos

fastgps

histos

depth

haulout

sst WC tag data files downloaded via download_data will be stored in separate, tag-specific subdirectories. path2data should point to the outer directory. The function always chooses the latest WC-processed -Locations.csv & FastGPS.csv files.

pull_data(path2data, source = "smru", cids = NULL, tag_mfr = "smru", ...)

Arguments

path2data

path to data file(s)

source

one of "smru", "wc", or "local"

cids

SMRU campaign ids. If not specified then the cids are built from the directory or filenames present in the path2data directory.

tag_mfr

either "smru or "wc", ignored if source != "local"

...

additional arguments passed to pull_smru_tables