clean, truncate & convert diag files to sf-tibbles

restructures diag files, formats dates & lc's; truncates start (and end for "nrt") of individual deployments using ctd dates; converts to sf geometry - all in preparation for SSM-filtering. Splits resulting truncated diag files by species.

smru_prep_loc(
  smru,
  meta,
  dropIDs = NULL,
  crs = NULL,
  extent = NULL,
  QCmode = NULL
)

Arguments

smru

list of SMRU tables

meta

metadata used to truncate start of diag data for each individual

dropIDs

SMRU refs to be dropped (eg. tags were turned on but not deployed)

crs

a proj4string to re-project diag locations from longlat. Default is NULL which results in one of 4 possible projections applied automatically, based on the centroid of the tracks. See overview vignette for details.

extent

geographic extent of locations, in longlat, to consider in QC

QCmode

specify whether QC is near real-time (nrt) or delayed-mode (dm), in latter case diag is not right-truncated & date of first dive is used for the track start date