metadata table tests & write to .csv

Apply AODN tests to metadata table, write to .csv - format depends on program (IMOS, ATN)

smru_write_meta(
  meta,
  program = "imos",
  test = TRUE,
  path = NULL,
  dropIDs = NULL,
  suffix = "_nrt"
)

Arguments

meta

metadata

program

Determines structure of output metadata. Currently, either imos or atn.

test

should variables be tested for standards compliance, default is TRUE. Standards compliance is specific to the program. Currently, only program = imos has defined variable standard against which output compliance is tested.

path

path to write .csv files

dropIDs

individual ids to be dropped

suffix

suffix to add to .csv files (_nrt, _dm, or _hist)