WC API access to tag data files

Download satellite tracking data and associated tag files via the Wildlife Computers Portal API. User-supplied WC Access Key, WC Secret Key,

Satellite tracking data are downloaded from a user-specified URL (which can include files on GoogleDrive or Dropbox), accessed from the SMRU data server by calling get_smru_mdb(), or accessed from the Wildlife Computers Portal API by calling get_wc_files().

wc_get_files(
  dest = NULL,
  a.key = NULL,
  s.key = NULL,
  owner.id = NULL,
  subset.ids = NULL,
  collaborator = TRUE,
  unzip = TRUE,
  verbose = FALSE,
  download = TRUE,
  return.tag.meta = FALSE
)

Arguments

a.key

an Access Key issued by Wildlife Computers for their API

s.key

a Secret Key issued by Wildlife Computers for their API

owner.id

a WC data owner ID

subset.ids

a single column .CSV file of WC UUID's to be included in the QC, with uuid as the variable name.

collaborator

(logical) should data owned by collaborators be downloaded. Ignored if owner.id provided.

unzip

(logical) should deployment zipfile be unzipped into destination directory.

download

(logical) should the tag data files be download from the WC Data portal. If FALSE then only the WC tag metadata is downloaded.

return.tag.meta

(logical) should tag metadata constructed from WC dataset information be returned